eslint errors fixed
