+++ src/ui.js
@@ -1,4 +1,5 @@
 import { IconPicture } from '@codexteam/icons';
+import { make } from './utils/dom';
 
 /**
  * Class for working with UI:
@@ -250,26 +251,3 @@ export default class Ui {
   }
 }
 
-/**
- * Helper for making Elements with attributes
- *
- * @param  {string} tagName           - new Element tag name
- * @param  {Array|string} classNames  - list or name of CSS class
- * @param  {object} attributes        - any attributes
- * @returns {Element}
- */
-export const make = function make(tagName, classNames = null, attributes = {}) {
-  const el = document.createElement(tagName);
-
-  if (Array.isArray(classNames)) {
-    el.classList.add(...classNames);
-  } else if (classNames) {
-    el.classList.add(classNames);
-  }
-
-  for (const attrName in attributes) {
-    el[attrName] = attributes[attrName];
-  }
-
-  return el;
-};
